+++ cypress/integration/allEnv/configuration/channels.js
@@ -14,8 +14,8 @@ import { CHANNELS_SELECTORS } from "../../../elements/channels/channels-selector
 import { SELECT_CHANNELS_TO_ASSIGN } from "../../../elements/channels/select-channels-to-assign";
 import { HEADER_SELECTORS } from "../../../elements/header/header-selectors";
 import { BUTTON_SELECTORS } from "../../../elements/shared/button-selectors";
-import { SHARED_ELEMENTS } from "../../../elements/shared/sharedElements";
 import { createChannelByView } from "../../../steps/channelsSteps";
+import { waitForProgressBarToNotExist } from "../../../steps/shared/progressBar";
 import { urlList } from "../../../url/urlList";
 import { deleteChannelsStartsWith } from "../../../utils/channelsUtils";
 import { deleteShippingStartsWith } from "../../../utils/shippingUtils";
@@ -46,6 +46,7 @@ describe("Channels", () => {
     const randomChannel = `${channelStartsWith} ${faker.datatype.number()}`;
     cy.addAliasToGraphRequest("Channels");
     cy.visit(urlList.channels);
+    cy.softExpectSkeletonIsVisible();
     cy.wait("@Channels");
     createChannelByView({ name: randomChannel, currency });
     cy.wait("@Channel");
@@ -69,7 +70,7 @@ describe("Channels", () => {
     cy.addAliasToGraphRequest("InitialProductFilterAttributes");
     cy.visit(urlList.products);
     cy.wait("@InitialProductFilterAttributes");
-    cy.get(SHARED_ELEMENTS.progressBar).should("not.exist");
+    waitForProgressBarToNotExist();
     cy.get(PRODUCTS_LIST.emptyProductRow).should("not.exist");
     cy.get(PRODUCTS_LIST.productsList)
       .first()
@@ -87,6 +88,7 @@ describe("Channels", () => {
     const randomChannel = `${channelStartsWith} ${faker.datatype.number()}`;
     cy.addAliasToGraphRequest("Channels");
     cy.visit(urlList.channels);
+    cy.softExpectSkeletonIsVisible();
     cy.wait("@Channels");
     createChannelByView({
       name: randomChannel,
@@ -111,6 +113,7 @@ describe("Channels", () => {
       currencyCode: currency
     });
     cy.visit(urlList.channels);
+    cy.softExpectSkeletonIsVisible();
     createChannelByView({ name: randomChannel, currency });
     cy.get(ADD_CHANNEL_FORM_SELECTORS.slugValidationMessage).should(
       "be.visible"
@@ -120,6 +123,7 @@ describe("Channels", () => {
   it("should validate duplicated currency", () => {
     const randomChannel = `${channelStartsWith} ${faker.datatype.number()}`;
     cy.visit(urlList.channels);
+    cy.softExpectSkeletonIsVisible();
     createChannelByView({
       name: randomChannel,
       currency: "notExistingCurrency"
@@ -139,6 +143,7 @@ describe("Channels", () => {
     });
     cy.addAliasToGraphRequest("Channels");
     cy.visit(urlList.channels);
+    cy.softExpectSkeletonIsVisible();
     cy.wait("@Channels");
     cy.contains(CHANNELS_SELECTORS.channelName, randomChannelToDelete)
       .parentsUntil(CHANNELS_SELECTORS.channelsTable)