+++ cypress/steps/catalog/products/VariantsSteps.js
@@ -3,8 +3,8 @@ import { PRODUCT_DETAILS } from "../../../elements/catalog/products/product-deta
 import { VARIANTS_SELECTORS } from "../../../elements/catalog/products/variants-selectors";
 import { AVAILABLE_CHANNELS_FORM } from "../../../elements/channels/available-channels-form";
 import { BUTTON_SELECTORS } from "../../../elements/shared/button-selectors";
-import { SHARED_ELEMENTS } from "../../../elements/shared/sharedElements";
 import { selectChannelVariantInDetailsPage } from "../../channelsSteps";
+import { waitForProgressBarToNotBeVisible } from "../../shared/progressBar";
 import { fillUpPriceList } from "./priceList";
 
 export function variantsShouldBeVisible({ name, price }) {
@@ -32,8 +32,7 @@ export function createFirstVariant({ sku, warehouseId, price, attribute }) {
   cy.addAliasToGraphRequest("ProductVariantBulkCreate");
   cy.get(VARIANTS_SELECTORS.nextButton).click();
   cy.wait("@ProductVariantBulkCreate");
-  cy.get(SHARED_ELEMENTS.progressBar)
-    .should("not.be.visible")
+  waitForProgressBarToNotBeVisible()
     .get(AVAILABLE_CHANNELS_FORM.menageChannelsButton)
     .should("be.visible");
 }
@@ -71,10 +70,8 @@ export function createVariant({
   cy.addAliasToGraphRequest("ProductVariantChannelListingUpdate");
   cy.get(VARIANTS_SELECTORS.saveButton).click();
   cy.wait("@ProductVariantChannelListingUpdate");
-  cy.get(BUTTON_SELECTORS.back)
-    .click()
-    .get(SHARED_ELEMENTS.progressBar)
-    .should("not.be.visible")
+  cy.get(BUTTON_SELECTORS.back).click();
+  waitForProgressBarToNotBeVisible()
     .get(AVAILABLE_CHANNELS_FORM.menageChannelsButton)
     .should("be.visible");
 }