+++ cypress/url/urlList.js
@@ -16,9 +16,11 @@ export const urlList = {
   newPassword: "new-password/",
   permissionsGroups: "permission-groups/",
   categories: "categories/",
+  weightRete: "weight/",
   attributes: "attributes/",
   productTypes: "product-types/"
 };
+
 export const productDetailsUrl = productId => `${urlList.products}${productId}`;
 
 export const userDetailsUrl = userId => `${urlList.staffMembers}${userId}`;
@@ -34,6 +36,9 @@ export const categoryDetails = categoryId =>
 
 export const shippingZoneDetailsUrl = shippingZoneId =>
   `${urlList.shippingMethods}${shippingZoneId}`;
+
+export const weightRateUrl = (shippingZoneId, weightRateId) =>
+  `${urlList.shippingMethods}${shippingZoneId}/${urlList.weightRete}${weightRateId}`;
 
 export const warehouseDetailsUrl = warehouseId =>
   `${urlList.warehouses}${warehouseId}`;