+++ cypress/utils/products/productsUtils.js
@@ -22,7 +22,8 @@ export function createProductInChannel({
   visibleInListings = true,
   collectionId = null,
   description = null,
-  trackInventory = true
+  trackInventory = true,
+  weight = 1
 }) {
   let product;
   let variantsList;
@@ -54,7 +55,8 @@ export function createProductInChannel({
         quantityInWarehouse,
         channelId,
         price,
-        trackInventory
+        trackInventory,
+        weight
       });
     })
     .then(variantsResp => {